requirements.txt
django
python-docx
html2text
markdown